Update Group_03_Analysis.qmd
This is a new version of Group_03_Analysis.qmd, including Data Cleaning, EDA and a simple Model Fitting. In next step: @yixinning nedd to add the tables of summary statistics in the EDA part, @kakiya-428 need to add the Heat maps for numerical variables in the EDA part and @hansomeLiu need to supplement the Model Fitting part.

Co-Authored-By: yixinning <202712693+yixinning@users.noreply.github.com>
Co-Authored-By: Heng YANG <187826578+kakiya-428@users.noreply.github.com>
Co-Authored-By: Haotian Liu <128010290+hansomeLiu@users.noreply.github.com>