Updated Group_03_Analysis.qmd
This is a new version of Group_03_Analysis.qmd, including Data Cleaning, EDA and a simple Model Fitting. In our next step: @yixinning need to add the tables of summary statistics in the EDA part, @kakiya-428 need to add the Heat maps for numerical variables in the EDA part and @hansomeLiu need to supplement the Model Fitting part.