Updated Group_03_Analysis.qmd
I have retained the heatmap for all numerical variables, added histograms for numerical variables, corrected the incorrectly labeled dependent variable in the first table, adjusted the order of the boxplots, converted House Age and Number of Bedrooms into categorical variables, and finally added a log transformation in the initial modeling section. In the next step, @yixinning can add the comment to the tables, @Chris-0806 can add Introduction part and add comment to the plots, and @hansomeLiu can supplement the model fitting part.